Add ability to respect z order (in addition to insertion order)
In the SVG standard (v1.1), it is said that the z order is determined by the insertion order of SVG elements. To simplify setting the z order, this commit adds a "z" element to specify it explicitly. This is used to sort the added elements according to z before writing the SVG XML code.